+++ .changeset/young-bulldogs-raise.md
@@ -1,0 +1,5 @@
+---
+"claude-dev": patch
+---
+
+Refactor task class, moving auto approve
